Tabs can be created and resorted, Icon selection dialog, Settings categories, Inactive icon states, Added an catalog icon, setOnApplyUiInsetsListener now allows propagation for children, Fixed crash when columns can't fit in the SearchActivity

+++ app/src/main/java/com/mrboomdev/awery/data/settings/SettingsItem.java
@@ -272,7 +272,7 @@ public class SettingsItem {
 
 	public void onClick(Context context) {}
 
-	public boolean onDragged(long fromPosition, long toPosition) {
+	public boolean onDragged(int fromPosition, int toPosition) {
 		return false;
 	}
